refactor: insert 패키지 생성하여 이동
